Split CQuery file into multiple files
